Change the whole image set for the setup
The previous USBC board image one was in my opinion confusing cause it either
looked like there was a connection of B8 with some unknown pin (possibly spare
GND, while correct, confusing for someone that just counts pins) or B7 with B12
that wasn't correct. The new picture has labels and leaves nothing to the
imagination.

And to be consistent the whole setup has been changed.